server/requirements.txt
sanic
sqlalchemy
aiomysql
pyJWT
bcrypt
python-decouple
jsonschema 
sanic-openapi
requests
##py -m pip install -r requirements.txt 